refactor(core): combine linters
