+++ harper-core/src/linting/mod.rs
@@ -6,7 +6,6 @@ mod a_part;
 mod addicting;
 mod adjective_double_degree;
 mod adjective_of_a;
-mod affect_effect;
 mod am_in_the_morning;
 mod amounts_for;
 mod an_a;
@@ -153,7 +152,6 @@ pub use a_part::APart;
 pub use addicting::Addicting;
 pub use adjective_double_degree::AdjectiveDoubleDegree;
 pub use adjective_of_a::AdjectiveOfA;
-pub use affect_effect::AffectEffect;
 pub use am_in_the_morning::AmInTheMorning;
 pub use amounts_for::AmountsFor;
 pub use an_a::AnA;
